requirements.txt
python-dotenv==0.21.0
Flask==2.2.2
Flask-RESTful==0.3.9
SQLAlchemy==1.4.40
Flask-SQLAlchemy==2.5.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
flask_marshmallow==0.14.0
asgiref==3.6.0
psycopg2-binary==2.9.3
flask-uploads==0.2.1
python-ffmpeg==2.0.4
pydub==0.25.1
